refs: keep track of unresolved reference value in iterator
Since ref iterators do not hold onto the value of a symbolic reference
before resolution, the only way to get ahold of a direct value of a
symbolic ref is to make a separate call to refs_read_symbolic_ref which
can be wasteful when iterating over all references.

To make accessing the direct value of a symbolic ref more efficient,
let's keep track of the value in the iterator so callbacks can have
access to it.

To do so, we need to add an argument to refs_resolve_ref_unsafe to save
the direct value of the reference to. Additionally, adjust
files-backend.c and reftable-backend.c to handle passing this value into
ther respective iterators.

Signed-off-by: John Cai <johncai86@gmail.com>