ref-filter: populate symref from iterator
With a previous commit, the reference the symbolic ref points is saved
in the ref iterator records. Instead of making a separate call to
resolve_refdup() each time, we can just populate the ref_array_item with
the value from the iterator.

This change results in a nice speedup in git-for-each-ref(1). When run
with about 1300 symbolic references, we see the following benchmark:

Benchmark 1: git for-each-ref

  Time (mean ± σ):      12.7 ms ±   0.7 ms    [User: 5.1 ms, System: 7.6 ms]
  Range (min … max):    11.9 ms …  17.0 ms    168 runs

Benchmark 2: ~/Projects/git/git for-each-ref

  Time (mean ± σ):       7.1 ms ±   0.5 ms    [User: 3.3 ms, System: 3.9 ms]
  Range (min … max):     6.6 ms …  10.7 ms    375 runs

Summary
  '~/Projects/git/git for-each-ref' ran
    1.79 ± 0.15 times faster than 'git for-each-ref'

Signed-off-by: John Cai <johncai86@gmail.com>